+++ wallet/api/errors.go
@@ -3,6 +3,7 @@ package api
 import (
 	"context"
 	"errors"
+	"fmt"
 
 	"code.vegaprotocol.io/vega/libs/jsonrpc"
 	"code.vegaprotocol.io/vega/wallet/api/node/types"
@@ -117,15 +118,10 @@ var (
 	ErrMultipleNetworkSources                             = errors.New("network sources are mutually exclusive")
 	ErrNetworkAlreadyExists                               = errors.New("a network with the same name already exists")
 	ErrNetworkConfigurationDoesNotHaveGRPCNodes           = errors.New("the network does not have gRPC hosts configured")
-	ErrNetworkCouldNotProcessTransaction                  = errors.New("the network could not process the transaction")
 	ErrNetworkDoesNotExist                                = errors.New("the network does not exist")
 	ErrNetworkNameIsRequired                              = errors.New("the network name is required")
 	ErrNetworkOrNodeAddressIsRequired                     = errors.New("a network or a node address is required")
-	ErrNetworkRejectedInvalidTransaction                  = errors.New("the network rejected the transaction because it's invalid")
-	ErrNetworkRejectedMalformedTransaction                = errors.New("the network rejected the transaction because it's malformed")
-	ErrNetworkRejectedUnsupportedTransaction              = errors.New("the network does not support this transaction")
 	ErrNetworkSourceIsRequired                            = errors.New("a network source is required")
-	ErrNetworkSpamProtectionActivated                     = errors.New("the network blocked the transaction through the spam protection")
 	ErrNewNameIsRequired                                  = errors.New("the new name is required")
 	ErrNewPassphraseIsRequired                            = errors.New("the new passphrase is required")
 	ErrNextAndCurrentPublicKeysCannotBeTheSame            = errors.New("the next and current public keys cannot be the same")
@@ -151,18 +147,23 @@ var (
 	ErrSpecifyingNetworkAndLastBlockDataIsNotSupported    = errors.New("specifying a network and the last block data is not supported")
 	ErrSpecifyingNetworkAndNodeAddressIsNotSupported      = errors.New("specifying a network and a node address is not supported")
 	ErrSubmissionBlockHeightIsRequired                    = errors.New("the submission block height is required")
+	ErrTransactionCouldNotBeSentThroughSelectedNode       = errors.New("the transaction could not be sent through the selected node")
 	ErrTransactionFailed                                  = errors.New("the transaction failed")
 	ErrTransactionIsNotValidJSON                          = errors.New("the transaction is not valid JSON")
 	ErrTransactionIsRequired                              = errors.New("the transaction is required")
 	ErrUserCanceledTheRequest                             = errors.New("the user canceled the request")
 	ErrUserCloseTheConnection                             = errors.New("the user closed the connection")
-	ErrUserRejectedTheRequest                             = errors.New("the user rejected the request")
+	ErrUserRejectedAccessToKeys                           = errors.New("the user rejected the access to the keys")
+	ErrUserRejectedSendingOfTransaction                   = errors.New("the user rejected the sending of the transaction")
+	ErrUserRejectedSigningOfTransaction                   = errors.New("the user rejected the signing of the transaction")
+	ErrUserRejectedWalletConnection                       = errors.New("the user rejected the wallet connection")
 	ErrWalletAlreadyExists                                = errors.New("a wallet with the same name already exists")
 	ErrWalletDoesNotExist                                 = errors.New("the wallet does not exist")
 	ErrWalletIsLocked                                     = errors.New("the wallet is locked")
 	ErrWalletIsRequired                                   = errors.New("the wallet is required")
 	ErrWalletKeyDerivationVersionIsRequired               = errors.New("the wallet key derivation version is required")
 	ErrWrongPassphrase                                    = errors.New("wrong passphrase")
+	ErrTransactionBlockedBySpamRules                      = errors.New("the transaction will break the network's spam rules")
 )
 
 func applicationError(code jsonrpc.ErrorCode, err error) *jsonrpc.ErrorDetails {
@@ -191,22 +192,22 @@ func networkErrorFromTransactionError(err error) *jsonrpc.ErrorDetails {
 	txErr := types.TransactionError{}
 	isTxErr := errors.As(err, &txErr)
 	if !isTxErr {
-		return networkError(ErrorCodeNodeCommunicationFailed, ErrTransactionFailed)
+		return networkError(ErrorCodeNodeCommunicationFailed, fmt.Errorf("the transaction failed: %w", err))
 	}
 
 	switch txErr.ABCICode {
 	case 51:
-		return networkError(ErrorCodeNetworkRejectedInvalidTransaction, ErrNetworkRejectedInvalidTransaction)
+		return networkError(ErrorCodeNetworkRejectedInvalidTransaction, fmt.Errorf("the network rejected the transaction because it's invalid: %w", err))
 	case 60:
-		return networkError(ErrorCodeNetworkRejectedMalformedTransaction, ErrNetworkRejectedMalformedTransaction)
+		return networkError(ErrorCodeNetworkRejectedMalformedTransaction, fmt.Errorf("the network rejected the transaction because it's malformed: %w", err))
 	case 70:
-		return networkError(ErrorCodeNetworkCouldNotProcessTransaction, ErrNetworkCouldNotProcessTransaction)
+		return networkError(ErrorCodeNetworkCouldNotProcessTransaction, fmt.Errorf("the network could not process the transaction: %w", err))
 	case 80:
-		return networkError(ErrorCodeNetworkRejectedUnsupportedTransaction, ErrNetworkRejectedUnsupportedTransaction)
+		return networkError(ErrorCodeNetworkRejectedUnsupportedTransaction, fmt.Errorf("the network does not support this transaction: %w", err))
 	case 89:
-		return networkError(ErrorCodeNetworkSpamProtectionActivated, ErrNetworkSpamProtectionActivated)
+		return networkError(ErrorCodeNetworkSpamProtectionActivated, fmt.Errorf("the network blocked the transaction through the spam protection: %w", err))
 	default:
-		return networkError(ErrorCodeNetworkRejectedTransaction, ErrTransactionFailed)
+		return networkError(ErrorCodeNetworkRejectedTransaction, fmt.Errorf("the transaction failed: %w", err))
 	}
 }
 
@@ -234,8 +235,8 @@ func userCancellationError(err error) *jsonrpc.ErrorDetails {
 	return userError(ErrorCodeRequestHasBeenCanceledByUser, err)
 }
 
-func userRejectionError() *jsonrpc.ErrorDetails {
-	return userError(ErrorCodeRequestHasBeenRejected, ErrUserRejectedTheRequest)
+func userRejectionError(err error) *jsonrpc.ErrorDetails {
+	return userError(ErrorCodeRequestHasBeenRejected, err)
 }
 
 func applicationCancellationError(err error) *jsonrpc.ErrorDetails {